mezzo-v2/requirements.txt
# Основные зависимости
Python==3.11.2
Flas==3.1.1
Werkzeug==3.1.3
